+++ src/items.c
@@ -501,6 +501,7 @@ void item_serialize(gpointer oid, gpointer it, gpointer root)
     if (i->bonus_known == 1) cJSON_AddTrueToObject(ival, "bonus_known");
     if (i->fired == 1) cJSON_AddTrueToObject(ival, "fired");
     if (i->picked_up) cJSON_AddTrueToObject(ival, "picked_up");
+    if (i->player_owned) cJSON_AddTrueToObject(ival, "player_owned");
 
     if (i->corroded > 0) cJSON_AddNumberToObject(ival, "corroded", i->corroded);
     if (i->burnt > 0) cJSON_AddNumberToObject(ival, "burnt", i->burnt);
@@ -577,6 +578,9 @@ item *item_deserialize(cJSON *iser, struct game *g)
 
     obj = cJSON_GetObjectItem(iser, "picked_up");
     if (obj != NULL) it->picked_up = obj->valueint;
+
+    obj = cJSON_GetObjectItem(iser, "player_owned");
+    if (obj != NULL) it->player_owned = obj->valueint;
 
     obj = cJSON_GetObjectItem(iser, "corroded");
     if (obj != NULL) it->corroded = obj->valueint;